Fix canvas drawImage full snapshot overwriting incremental snapshot drawImage (#5)
## The Problem

rrweb-snapshot is supposed to process full snapshots synchronously.
However, canvas functionality was seemingly added after this design
decision. Canvas drawImage takes an image element as its first argument.
Snapshot rebuilding waits for this image to load before calling
drawImage. This creates a problem when subsequent incremental snapshots
call drawImage on the same canvas element. The full snapshot's drawImage
call can end up overwriting subsequent drawImage calls.

In session playback, we see this problem manifest as confusing flashes
back to the first frame of the session recording on operations like
pause or skip back because the full snapshot is getting processed again.

## The Solution

In this PR, we patch the canvas context's drawImage function to keep
track of pending image loads. If drawImage gets called with a new image
while a previous image is still loading, we cancel the pending drawImage
operation for the old image and let it get replaced by the new image.